refactor: change int -> long where necessary

+++ src/main/java/com/picpay/transfers/domain/TransferPayee.java
@@ -3,6 +3,6 @@ package com.picpay.transfers.domain;
 import com.picpay.transfers.domain.value_objects.Money;
 
 public interface TransferPayee {
-    int id();
-    void receiveTransfer(Money amount, int payerId, int payeeId);
+    long id();
+    void receiveTransfer(Money amount, long payerId, long payeeId);
 }
